Injest the new server files from bundle in app

+++ Canyoneer/CanyoneerTests/View/Canyon/StarQualityViewModelTests.swift
@@ -7,61 +7,61 @@ import XCTest
 @MainActor
 class StarQualityViewModelTests: XCTestCase {
     func testStars_zero() {
-        let quality: Float = 0
+        let quality: Double = 0
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 0)
     }
     
     func testStars_one() {
-        let quality: Float = 1
+        let quality: Double = 1
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 1)
     }
     
     func testStars_one_one() {
-        let quality: Float = 1.1
+        let quality: Double = 1.1
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 1)
     }
     
     func testStars_one_five() {
-        let quality: Float = 1.5
+        let quality: Double = 1.5
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 2)
     }
     
     func testStars_one_nine() {
-        let quality: Float = 1.9
+        let quality: Double = 1.9
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 2)
     }
     
     func testStars_two() {
-        let quality: Float = 2
+        let quality: Double = 2
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 2)
     }
     
     func testStars_two_five() {
-        let quality: Float = 2.5
+        let quality: Double = 2.5
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 3)
     }
     
     func testStars_three() {
-        let quality: Float = 3
+        let quality: Double = 3
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 3)
     }
     
     func testStars_four() {
-        let quality: Float = 4
+        let quality: Double = 4
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 4)
     }
     
     func testStars_five() {
-        let quality: Float = 5
+        let quality: Double = 5
         let result = StarQualityViewModel.stars(quality: quality)
         XCTAssertEqual(result.count, 5)
     }
